refactor: Rename jsonStore -> store, extracted handlers tests
